temporary fixed tests with just deleting user when test finishes
